tidy up names
